requirements.txt
asgiref==3.11.0
Django==5.2.8
djangorestframework==3.16.1
flake8==7.3.0
gunicorn==23.0.0
iniconfig==2.3.0
mccabe==0.7.0
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.11
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
pytest==9.0.1
pytest-django==4.11.1
python-dotenv==1.2.1
sqlparse==0.5.4
